Adjusted Graphwiz Syscall to new behaviour
Graphwiz now expects a SIGTERM when you dont need it anymore. Graflap didnt handle that so it ended in Graflap waiting for Graphwiz to stop while Graphwiz waited for a Signal to halt.
Added code that sends a SIGTERM after the task has been processed and the answer buffered in Graflap.